Created project module and moved Scheme into it
