+++ glotaran/analysis/test/test_problem.py
@@ -2,12 +2,15 @@ import collections
 
 import numpy as np
 import pytest
+import xarray as xr
 
 from glotaran.analysis.problem import Problem
-from glotaran.analysis.scheme import Scheme
 from glotaran.analysis.simulation import simulate
+from glotaran.parameter import ParameterGroup
+from glotaran.project import Scheme
 
 from .models import MultichannelMulticomponentDecay as suite
+from .models import SimpleTestModel
 
 
 @pytest.fixture(
@@ -149,3 +152,71 @@ def test_problem_result_data(problem: Problem):
 
     assert "residual_singular_values" in dataset
     assert "weighted_residual_singular_values" in dataset
+
+
+def test_prepare_data():
+    model_dict = {
+        "dataset": {
+            "dataset1": {
+                "megacomplex": [],
+            },
+        },
+        "weights": [
+            {
+                "datasets": ["dataset1"],
+                "global_interval": (np.inf, 200),
+                "model_interval": (4, 8),
+                "value": 0.5,
+            },
+        ],
+    }
+    model = SimpleTestModel.from_dict(model_dict)
+    print(model.validate())
+    assert model.valid()
+
+    parameters = ParameterGroup.from_list([])
+
+    global_axis = np.asarray(range(50, 300))
+    model_axis = np.asarray(range(15))
+
+    dataset = xr.DataArray(
+        np.ones((global_axis.size, model_axis.size)),
+        coords={"e": global_axis, "c": model_axis},
+        dims=("e", "c"),
+    )
+
+    scheme = Scheme(model, parameters, {"dataset1": dataset})
+    problem = Problem(scheme)
+
+    data = problem.data["dataset1"]
+    print(data)
+    assert "data" in data
+    assert "weight" in data
+
+    assert data.data.shape == (model_axis.size, global_axis.size)
+    assert data.data.shape == data.weight.shape
+    assert np.all(data.weight.sel(e=slice(0, 200), c=slice(4, 8)).values == 0.5)
+    assert np.all(data.weight.sel(c=slice(0, 3)).values == 1)
+
+    model_dict["weights"].append(
+        {
+            "datasets": ["dataset1"],
+            "value": 0.2,
+        }
+    )
+    model = SimpleTestModel.from_dict(model_dict)
+    print(model.validate())
+    assert model.valid()
+
+    scheme = Scheme(model, parameters, {"dataset1": dataset})
+    problem = Problem(scheme)
+    data = problem.data["dataset1"]
+    assert np.all(data.weight.sel(e=slice(0, 200), c=slice(4, 8)).values == 0.5 * 0.2)
+    assert np.all(data.weight.sel(c=slice(0, 3)).values == 0.2)
+
+    with pytest.warns(
+        UserWarning,
+        match="Ignoring model weight for dataset 'dataset1'"
+        " because weight is already supplied by dataset.",
+    ):
+        Problem(Scheme(model, parameters, {"dataset1": data}))
