+++ glotaran/builtin/io/yml/yml.py
@@ -1,10 +1,16 @@
+import pathlib
+
 import yaml
 
+from glotaran.io import load_model
+from glotaran.io import load_parameters
+from glotaran.io import read_data_file
 from glotaran.io import register_io
 from glotaran.model import Model
 from glotaran.model import get_model
 from glotaran.model import known_model
 from glotaran.parameter import ParameterGroup
+from glotaran.project import Scheme
 
 from .sanatize import sanitize_yaml
 
@@ -61,3 +67,73 @@ class YmlIo:
             return ParameterGroup.from_list(spec)
         else:
             return ParameterGroup.from_dict(spec)
+
+    def read_scheme(fmt: str, file_name: str) -> Scheme:
+        if fmt == "yml_str":
+            yml = file_name
+        else:
+            try:
+                with open(file_name) as f:
+                    yml = f.read()
+            except Exception as e:
+                raise OSError(f"Error opening scheme: {e}")
+
+        try:
+            scheme = yaml.safe_load(yml)
+        except Exception as e:
+            raise ValueError(f"Error parsing scheme: {e}")
+
+        if "model" not in scheme:
+            raise ValueError("Model file not specified.")
+
+        try:
+            model = load_model(scheme["model"])
+        except Exception as e:
+            raise ValueError(f"Error loading model: {e}")
+
+        if "parameters" not in scheme:
+            raise ValueError("Parameters file not specified.")
+
+        try:
+            parameters = load_parameters(scheme["parameters"])
+        except Exception as e:
+            raise ValueError(f"Error loading parameters: {e}")
+
+        if "data" not in scheme:
+            raise ValueError("No data specified.")
+
+        data = {}
+        for label, path in scheme["data"].items():
+            path = pathlib.Path(path)
+
+            fmt = path.suffix[1:] if path.suffix != "" else "nc"
+            if "data_format" in scheme:
+                fmt = scheme["data_format"]
+
+            try:
+                data[label] = read_data_file(path, fmt=fmt)
+            except Exception as e:
+                raise ValueError(f"Error loading dataset '{label}': {e}")
+
+        optimization_method = scheme.get("optimization_method", "TrustRegionReflection")
+        nnls = scheme.get("non-negative-least-squares", False)
+        nfev = scheme.get("maximum-number-function-evaluations", None)
+        ftol = scheme.get("ftol", 1e-8)
+        gtol = scheme.get("gtol", 1e-8)
+        xtol = scheme.get("xtol", 1e-8)
+        group_tolerance = scheme.get("group_tolerance", 0.0)
+        return Scheme(
+            model=model,
+            parameters=parameters,
+            data=data,
+            non_negative_least_squares=nnls,
+            maximum_number_function_evaluations=nfev,
+            ftol=ftol,
+            gtol=gtol,
+            xtol=xtol,
+            group_tolerance=group_tolerance,
+            optimization_method=optimization_method,
+        )
+
+    def write_scheme(fmt: str, file_name: str, result: Scheme):
+        raise NotImplementedError
