+++ glotaran/builtin/models/kinetic_spectrum/test/test_coherent_artifact.py
@@ -2,12 +2,12 @@ import numpy as np
 import xarray as xr
 
 from glotaran.analysis.optimize import optimize
-from glotaran.analysis.scheme import Scheme
 from glotaran.builtin.models.kinetic_spectrum import KineticSpectrumModel
 from glotaran.builtin.models.kinetic_spectrum.kinetic_spectrum_matrix import (
     kinetic_spectrum_matrix,
 )
 from glotaran.parameter import ParameterGroup
+from glotaran.project import Scheme
 
 
 def test_coherent_artifact():
