+++ glotaran/builtin/models/kinetic_spectrum/test/test_spectral_constraints.py
@@ -3,13 +3,13 @@ import pytest
 import xarray as xr
 
 from glotaran.analysis.optimize import optimize
-from glotaran.analysis.scheme import Scheme
 from glotaran.builtin.models.kinetic_image.kinetic_image_matrix import kinetic_image_matrix
 from glotaran.builtin.models.kinetic_spectrum import KineticSpectrumModel
 from glotaran.builtin.models.kinetic_spectrum.spectral_constraints import (
     apply_spectral_constraints,
 )
 from glotaran.parameter import ParameterGroup
+from glotaran.project import Scheme
 
 
 def test_spectral_constraint():
