+++ glotaran/builtin/models/kinetic_spectrum/test/test_spectral_irf.py
@@ -2,9 +2,9 @@ import numpy as np
 import pytest
 
 from glotaran.analysis.optimize import optimize
-from glotaran.analysis.scheme import Scheme
 from glotaran.io import load_model
 from glotaran.io import load_parameters
+from glotaran.project import Scheme
 
 MODEL_BASE = """\
 type: kinetic-spectrum
