+++ glotaran/builtin/models/kinetic_spectrum/test/test_spectral_relations.py
@@ -2,13 +2,13 @@ import numpy as np
 import xarray as xr
 
 from glotaran.analysis.optimize import optimize
-from glotaran.analysis.scheme import Scheme
 from glotaran.builtin.models.kinetic_image.kinetic_image_matrix import kinetic_image_matrix
 from glotaran.builtin.models.kinetic_spectrum import KineticSpectrumModel
 from glotaran.builtin.models.kinetic_spectrum.spectral_relations import (
     create_spectral_relation_matrix,
 )
 from glotaran.parameter import ParameterGroup
+from glotaran.project import Scheme
 
 
 def test_spectral_relation():
