+++ glotaran/io/__init__.py
@@ -6,6 +6,7 @@ from .decorator import register_io
 from .io import Io
 from .register import load_model
 from .register import load_parameters
+from .register import load_scheme
 from .register import write_parameters
 
 prepare_time_trace_dataset = prepare_dataset.prepare_time_trace_dataset
